Refactor program to use only 1 form
Refactored the UX and GUI in how you no longer need to click on a list to then edit it's tedious value, you can now just edit it in the form. Also formatted and cleaned up the code and updated it for .NET 10 and got it optimized and shortened and simplified the code to prepare for the Avalonia UI port Update based on this new commit.